+++ data_structures/VertexEdge.cpp
@@ -8,11 +8,13 @@ Vertex::Vertex() {
 
 }
 
+
+
 /*
  * Auxiliary function to add an outgoing edge to a vertex (this),
  * with a given destination vertex (d) and edge weight (w).
  */
-Edge * Vertex::addEdge(Vertex *d, double w) {
+Edge * Vertex::addEdge(Vertex *d, int w) {
     auto newEdge = new Edge(this, d, w);
     adj.push_back(newEdge);
     d->incoming.push_back(newEdge);
@@ -132,7 +134,7 @@ Station Vertex::getStation() const {
 
 /********************** Edge  ****************************/
 
-Edge::Edge(Vertex *orig, Vertex *dest, double w): orig(orig), dest(dest), weight(w) {}
+Edge::Edge(Vertex *orig, Vertex *dest, int w): orig(orig), dest(dest), weight(w) {}
 Edge::Edge(){
 
 }
